app/requirements.txt
mysql-connector-python
Flask
watchdog
openpyxl
pandas
Flask-Migrate
flask-sqlalchemy
pymysql
cryptography